requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.9
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
python-multipart==0.0.6
ta-lib==0.4.28
scikit-learn==1.3.2
yfinance==0.2.18
plotly==5.17.0
asyncio==3.4.3
websockets==11.0.3
python-dateutil==2.8.2
pytz==2023.3 